Change get_time_ms return type to double
Include sub-millisecond precision by using a double and dividing the
microseconds by 1000.0.

+++ src/table/table.h
@@ -52,7 +52,7 @@ void						table_destroy(t_table *table);
 void						stop_simulation(t_table *table);
 bool						is_stop(t_table *table);
 
-long						get_time_ms(void);
+double						get_time_ms(void);
 int							sleep_coder_ms(t_coder *coder, long sleep_ms);
 
 #endif
