+++ src/table/time.c
@@ -14,12 +14,12 @@
 #include "sys/time.h"
 #include "table.h"
 
-long	get_time_ms(void)
+double	get_time_ms(void)
 {
 	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return (tv.tv_sec * 1000 + tv.tv_usec / 1000);
+	return (tv.tv_sec * 1000 + tv.tv_usec / 1000.0);
 }
 
 int	sleep_coder_ms(t_coder *coder, long sleep_ms)
